+++ src/webapp-core/export/BaseApplication.js
@@ -19,19 +19,22 @@ export class BaseApplication extends BaseDispatched {
 
   /** @override */
   async handleRequest(request, dispatch) {
-    const logger    = this._prot_newDispatchLogger();
-    const requestId = logger ? request.id : null;
+    const logger    = this._prot_newDispatchLogger(dispatch.logger ?? request.id);
     const startTime = logger?.$env.now();
+
+    if (logger) {
+      dispatch = dispatch.withLogger(logger);
+    }
 
     const logDone = (fate, ...error) => {
       if (logger) {
         const endTime  = logger.$env.now();
         const duration = endTime.subtract(startTime);
-        logger[fate](requestId, duration, ...error);
+        logger[fate](duration, ...error);
       }
     };
 
-    logger?.handling(requestId, dispatch.infoForLog);
+    logger?.handling(dispatch.infoForLog);
 
     try {
       const result = await this.#callHandler(request, dispatch);